+++ src/cmd/help/index.ts
@@ -1,32 +1,38 @@
-import { AckFn, Block, KnownBlock, RespondArguments, SlashCommand } from "@slack/bolt";
+import {
+  AckFn,
+  Block,
+  KnownBlock,
+  RespondArguments,
+  SlashCommand,
+} from "@slack/bolt";
 import { ICommand } from "../interface";
-import { error, ezpr, overview, usage } from "./blocks";
+import { error, ezprHelp, helpOverview, helpUsage } from "./blocks";
 
 export class HelpCommand implements ICommand {
-  #ack: AckFn<string | RespondArguments>;
-  #message: (KnownBlock | Block)[];
+  ack: AckFn<string | RespondArguments>;
+  message: (KnownBlock | Block)[];
 
   constructor(ack: AckFn<string | RespondArguments>, payload: SlashCommand) {
-    this.#ack = ack;
+    this.ack = ack;
 
     switch (payload.text) {
       case "":
-        this.#message = overview
-        break
+        this.message = helpOverview;
+        break;
       case "usage":
-        this.#message = usage
-        break
+        this.message = helpUsage;
+        break;
       case "ezpr":
-        this.#message = ezpr
-        break
+        this.message = ezprHelp;
+        break;
       default:
-        this.#message = error(payload.text)
+        this.message = error(payload.text);
     }
   }
 
   async handle() {
-    await this.#ack({
-      blocks: this.#message,
+    await this.ack({
+      blocks: this.message,
       response_type: "ephemeral",
     });
   }