+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
-## V[X.Y.Z] - 2023-MM-DD
+## V[0.1.0] - 2023-01-01
  
+### Added
 
+- Help command
+    - `/help` to view an overview of EZ PR Bot with quick shortcuts
+    - `/help usage` to view a detailed overview of all the commands available
+    - `/help <command>` to view more information about a specific command